Redesign of the Back-End and Front-End. Significant improvements in API management and overall enhancement of the interface.
I improved the back-end but focused heavily on the front-end to normalize everything we had done. To achieve this, I started from a clean project and reimplemented what we had previously done to accelerate the redesign process.

During this process, I adapted and reviewed a large number of methods that were no longer suitable for the new version. I also added new features and included more comments to enhance code understanding.